Initial Commit: Git practice for team
This is a MATLAB file where each member is meant to add a line that prints their name.

Each member should pull the repository to their local repository, then they should change the code to print their name (as well as previous members names)

Then they should write a commit message explaining the changes to the file and commit the file. After commiting, the local repository should be pushed to the remote repository